Refactor canCreateChallenge typing
